Level-5,6: Handle errors and delete
Added try catch blocks and specific messages for each exception. Added the delete function in this commit as well, as it was just one extra method, and I was already using ArrayList to store tasks from the very beginning. Rewrote and shuffled around code in the main class to handle all the various exceptions. With how late I started this I am happy with my progress. I should add more flavour and personality in the upcoming weeks.